Remove unneeded steps
 * If cost is from a duplicate coordinate, means that such cost won't be at the
top of the heap. At least no while the last cost m-a, n-a indices was not calculated.
This can be done onñuy because I check if last element is set, which means that once
is set, it's impossible to get a better value for it, given that the cost from any
adjacent node to it is always the same. ONce we find it, it's the best value it can get